Add staff list to edit course
